Refactor: login function separation
